requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-camel-case2==0.3.0
Markdown==3.2.2
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
